Landscape_Gardening_Platform/requirements.txt
Django==4.2.18
djangorestframework==3.15.2
